Move types (stable)
