+++ apps/runner/connery-runner.config.ts
@@ -6,7 +6,7 @@
 // The .env file is not committed to the repository for security reasons.
 // ##################
 
-import { Config } from ':src/api/config/types';
+import { Config } from ':src/types/config-types';
 
 const config: Config = {
   runnerConfig: {
