+++ apps/runner/src/lib/__tests__/parameter-utils.test.ts
@@ -1,11 +1,11 @@
 import {
   InputParameterDefinition,
-  InputParametersObject,
   OutputParameterDefinition,
-  OutputParametersObject,
   ConfigurationParameterDefinition,
-  ConfigurationParametersObject,
-} from ':src/sdk/types';
+} from ':src/types/definition';
+import { ConfigurationParametersObject } from ':src/types/context';
+import { OutputParametersObject } from ':src/types/context';
+import { InputParametersObject } from ':src/types/context';
 import {
   trimInput,
   validateConfigurationParameterTypes,
